updated CLI structure
